Bagshui 1.4.7 (#122)
* Avoid renameGroup_Data error

The table being used to pass data for renaming Groups was sometimes getting removed. Make sure it stays.

* Fix Structure profiles out of sync at first login

When a character first logged in, the Bags and Bank Structure would end up pointing to different tables behind the scenes, causing confusion if any editing was done prior to a UI reload/logout. Fixes #121.

* Fix tooltip sizing in Edit Mode with left-anchored inventory

When an inventory window was anchored to the left (Bank default) and the cursor was moved from a group move target (the [+] widgets) to an item button, the tooltip would stretch all the way to the right of the screen.